Update HW_44.cs
изменила тип на дабл

+++ HW_44.cs
@@ -2,16 +2,16 @@
 // y = k1 * x + b1, y = k2 * x + b2, если b1, k1, b2 и k2 заданы
 
 Console.WriteLine("Введите b1");
-int b1 = int.Parse(Console.ReadLine());
+double b1 = int.Parse(Console.ReadLine());
 
 Console.WriteLine("Введите k1");
-int k1 = int.Parse(Console.ReadLine());
+double k1 = int.Parse(Console.ReadLine());
 
 Console.WriteLine("Введите b2");
-int b2 = int.Parse(Console.ReadLine());
+double b2 = int.Parse(Console.ReadLine());
 
 Console.WriteLine("Введите k2");
-int k2 = int.Parse(Console.ReadLine());
+double k2 = int.Parse(Console.ReadLine());
 
 if (k1 == k2)
 {
